refactor(project-sidebar): split use-project-sidebar into focused hooks (#973)
