+++ apps/web/lib/email/email-notification-service.tsx
@@ -1,6 +1,6 @@
-export { sendNewProjectEmails } from './notifications/send-new-project-emails'
-export { sendMilestoneApprovedNotification } from './notifications/send-milestone-approved'
-export { sendFundsReleasedNotification } from './notifications/send-funds-released'
-export { sendWelcomeEmail } from './notifications/send-welcome-email'
-export { sendContributionNotifications } from './notifications/send-contribution-notifications'
 export { sendCampaignGoalReachedNotifications } from './notifications/send-campaign-goal-reached'
+export { sendContributionNotifications } from './notifications/send-contribution-notifications'
+export { sendFundsReleasedNotification } from './notifications/send-funds-released'
+export { sendMilestoneApprovedNotification } from './notifications/send-milestone-approved'
+export { sendNewProjectEmails } from './notifications/send-new-project-emails'
+export { sendWelcomeEmail } from './notifications/send-welcome-email'
