+++ apps/web/lib/email/notifications/send-new-project-emails.tsx
@@ -1,5 +1,4 @@
 import { supabase as supabaseServiceRole } from '@packages/lib/supabase'
-import { NewProjectEmail } from '../templates/new-project-email'
 import {
 	appUrl,
 	createInAppNotification,
@@ -7,6 +6,7 @@ import {
 	getUserEmailPref,
 	sendEmail,
 } from '../notification-helpers'
+import { NewProjectEmail } from '../templates/new-project-email'
 
 export async function sendNewProjectEmails({
 	projectTitle,
